# internal/adapters/middleware/rest/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "IdPD identity provider server API.",
        "title": "IdPD REST API",
        "contact": {
            "url": "https://github.com/tdrn-org/idpd"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9123",
    "basePath": "/api",
    "paths": {
        "/api/info": {
            "get": {
                "description": "Query static server info like version and configured options",
                "produces": [
                    "application/json"
                ],
                "summary": "Query server info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ServerInfoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/ping": {
            "get": {
                "description": "Ping the server to check general health",
                "produces": [
                    "application/json"
                ],
                "summary": "Ping server",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.StatusResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/session": {
            "get": {
                "description": "Get the current session (if a session exists)",
                "produces": [
                    "application/json"
                ],
                "summary": "Get current session",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.SessionInfoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.SessionInfoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Initiate the authentication flow to create a new session",
                "produces": [
                    "application/json"
                ],
                "summary": "Initiate a new session",
                "responses": {
                    "302": {
                        "description": "Redirect to ...",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete the current session (if a session exists)",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete current session",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.StatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.StatusResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/session/login/{id}": {
            "get": {
                "description": "Get the login information for the authentication flow associated with the given authentication request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get login information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.SessionLoginInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.SessionLoginInfoResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/rest.SessionLoginInfoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Initiate the authentication flow to create a new session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.SessionLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.StatusResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/rest.StatusResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/session/verify": {
            "get": {
                "description": "Get the verify information for the authentication flow associated with the given authentication request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get verify information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.SessionVerifyInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Verify the authentication by providing the verification code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Verify authentication",
                "parameters": [
                    {
                        "description": "Request parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.SessionVerifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Redirect",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Verification": {
            "type": "string",
            "enum": [
                "email",
                "totp",
                "passkey",
                "seckey"
            ],
            "x-enum-varnames": [
                "VerificationEmail",
                "VerificationTOTP",
                "VerificationPasskey",
                "VerificationSecKey"
            ]
        },
        "rest.ServerInfo": {
            "type": "object",
            "properties": {
                "base_url": {
                    "description": "The server's base URL",
                    "type": "string"
                },
                "version": {
                    "description": "The server version",
                    "type": "string"
                }
            }
        },
        "rest.ServerInfoResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/rest.StatusCode"
                },
                "data": {
                    "$ref": "#/definitions/rest.ServerInfo"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "rest.SessionInfo": {
            "type": "object",
            "properties": {
                "strong_auth": {
                    "type": "boolean"
                },
                "user": {
                    "$ref": "#/definitions/rest.UserInfo"
                }
            }
        },
        "rest.SessionInfoResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/rest.StatusCode"
                },
                "data": {
                    "$ref": "#/definitions/rest.SessionInfo"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "rest.SessionLoginInfo": {
            "type": "object",
            "properties": {
                "allowed_verifications": {
                    "description": "AllowedVerifications lists the allowed verification methods for this login flow.\nOnly verification methods from this list are accepted during this login flow.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Verification"
                    }
                },
                "login_hint": {
                    "description": "LoginHint contains a hint for the login to use, if any can be derived from context.\nCan be overriden by the user.",
                    "type": "string"
                },
                "remember": {
                    "description": "Remember indicates the default for whether to remember the login across browser sessions or not.\nCan be overriden by the user.",
                    "type": "boolean"
                }
            }
        },
        "rest.SessionLoginInfoResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/rest.StatusCode"
                },
                "data": {
                    "$ref": "#/definitions/rest.SessionLoginInfo"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "rest.SessionLoginRequest": {
            "type": "object",
            "properties": {
                "login": {
                    "description": "Login is the user login to use for the authentication",
                    "type": "string"
                },
                "password": {
                    "description": "Password is the user password to use for the authentication",
                    "type": "string"
                },
                "remember": {
                    "description": "Remember indicates whether to remember the given login across browser sessions or not.",
                    "type": "boolean"
                },
                "verification": {
                    "description": "Verification is the verification method to perform in the next step auf the authentication flow.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Verification"
                        }
                    ]
                }
            }
        },
        "rest.SessionVerifyInfo": {
            "type": "object",
            "properties": {
                "verification": {
                    "$ref": "#/definitions/domain.Verification"
                }
            }
        },
        "rest.SessionVerifyRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code is the verification code",
                    "type": "string"
                },
                "id": {
                    "description": "ID identifies the authentication request this request refers to",
                    "type": "string"
                }
            }
        },
        "rest.StatusCode": {
            "type": "string",
            "enum": [
                "no_session",
                "auth_request_not_accessible"
            ],
            "x-enum-varnames": [
                "StatusCodeNoSession",
                "StatusCodeAuthRequestNotAccessible"
            ]
        },
        "rest.StatusResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/rest.StatusCode"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "rest.UserInfo": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "login": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                }
            }
        }
    }
}